move primus into the src subdir
Move primus dir into src, so we can get rid of babel-register on
production (very slow, not recommended). Also commented out primus
Multiplex (since it’s not used)
