+++ src/shared/data-updater.js
@@ -1,7 +1,7 @@
 
 export const DataUpdater = (playerName, type, data) => {
   // Would initialise server with testing if imported on top.
-  const primus = require('../../primus/server').primus;
+  const primus = require('../primus/server').primus;
 
   primus.forEach(spark => {
     if(!spark.authToken || spark.authToken.playerName !== playerName) return;
